5.4的劳动result

+++ src/main/java/njwb/lcqjoyce/bbs/mapper/UserMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface UserMapper {
     int deleteByPrimaryKey(Long userId);
 
-    int insert(User record);
+    long  insert(User record);
 
     int insertSelective(User record);
